src/e3v3se_display/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
multitimer==0.3
pyserial==3.5
requests==2.31.0
urllib3==2.1.0
